Removed Customizer as requirement.  Changed to customizer.yaml and customize_glob.yaml.
Removed Customizer as a requirement.  Attributes to be hidden are now set in customize.yaml and customize_glob.yaml. If `hide_attributes' is set in YAML, the `hide attributes` node will automatically be hidden, does not need to be explicitly defined.  `all` is a valid option to hide all attributes.

Example (customize.yaml):
```
input_boolean.development:
  hide_attributes:
    - editable
``

Example (customize_glob.yaml):
```
"*.*":
  hide_attributes:
    - all
``